+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.1.1 (26.11.2015)
 
 * #2: Do not exec the crontab, if the content has not changed
+* #4: Crontab::save() returns TRUE if the crontab has been modified
 
 ### 0.1.0 (18.11.2015)